# templates/golang/{{cookiecutter.project_name}}/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "蓝鲸开发者中心 SaaS 后台 API 文档",
        "title": "BkApp API DOC",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "crud"
                ],
                "summary": "获取分类列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serializer.CategoryListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "crud"
                ],
                "summary": "创建分类",
                "parameters": [
                    {
                        "description": "创建分类请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.CategoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "crud"
                ],
                "summary": "获取单个分类",
                "operationId": "get_category_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serializer.CategoryRetrieveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "crud"
                ],
                "summary": "更新分类",
                "operationId": "update_category_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新分类请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.CategoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "crud"
                ],
                "summary": "删除分类",
                "operationId": "delete_category_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/periodic-tasks": {
            "get": {
                "tags": [
                    "async-task"
                ],
                "summary": "获取定时任务列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serializer.PeriodicTaskListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "async-task"
                ],
                "summary": "创建定时任务",
                "parameters": [
                    {
                        "description": "定时任务配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.PeriodicTaskCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/periodic-tasks/{id}": {
            "delete": {
                "tags": [
                    "async-task"
                ],
                "summary": "删除定时任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "定时任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/periodic-tasks/{id}/enabled": {
            "put": {
                "tags": [
                    "async-task"
                ],
                "summary": "切换定时任务启用状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "定时任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ginx.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
...
        "serializer.CategoryCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                }
            }
        },
        "serializer.CategoryListResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "serializer.CategoryRetrieveResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "serializer.CategoryUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                }
            }
        },
...
        "serializer.PeriodicTaskCreateRequest": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {}
                },
                "cron": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "serializer.PeriodicTaskListResponse": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "cron": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}